+++ src/main/java/net/forcemaster_rpg/client/armor/AkenArmorRenderer.java
@@ -1,11 +1,20 @@
 package net.forcemaster_rpg.client.armor;
 
-import mod.azure.azurelibarmor.common.api.client.renderer.GeoArmorRenderer;
-import net.forcemaster_rpg.item.armor.AkenArmor;
-import net.forcemaster_rpg.item.armor.OrieneArmor;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRenderer;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRendererConfig;
+import net.minecraft.util.Identifier;
 
-public class AkenArmorRenderer extends GeoArmorRenderer<AkenArmor> {
-    public AkenArmorRenderer() {
-        super(new AkenArmorModel());
+import static net.forcemaster_rpg.ForcemasterClassMod.MOD_ID;
+
+public class AkenArmorRenderer extends AzArmorRenderer {
+    public static AkenArmorRenderer aken() {
+        return new AkenArmorRenderer("aken_armor", "aken_armor");
+    }
+
+    public AkenArmorRenderer(String modelName, String textureName) {
+        super(AzArmorRendererConfig.builder(
+                Identifier.of(MOD_ID, "geo/" + modelName + ".geo.json"),
+                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png")
+        ).build());
     }
 }