+++ src/main/java/net/forcemaster_rpg/client/armor/OrieneArmorRenderer.java
@@ -1,10 +1,20 @@
 package net.forcemaster_rpg.client.armor;
 
-import mod.azure.azurelibarmor.common.api.client.renderer.GeoArmorRenderer;
-import net.forcemaster_rpg.item.armor.OrieneArmor;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRenderer;
+import mod.azure.azurelibarmor.rewrite.render.armor.AzArmorRendererConfig;
+import net.minecraft.util.Identifier;
 
-public class OrieneArmorRenderer extends GeoArmorRenderer<OrieneArmor> {
-    public OrieneArmorRenderer() {
-        super(new OrieneArmorModel());
+import static net.forcemaster_rpg.ForcemasterClassMod.MOD_ID;
+
+public class OrieneArmorRenderer extends AzArmorRenderer {
+    public static OrieneArmorRenderer oriene() {
+        return new OrieneArmorRenderer("oriene_armor", "oriene_armor");
+    }
+
+    public OrieneArmorRenderer(String modelName, String textureName) {
+        super(AzArmorRendererConfig.builder(
+                Identifier.of(MOD_ID, "geo/" + modelName + ".geo.json"),
+                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png")
+        ).build());
     }
 }