+++ src/main/java/net/forcemaster_rpg/custom/custom_spells/CustomSpells.java
@@ -4,19 +4,18 @@ import net.forcemaster_rpg.effect.Effects;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.Vec3d;
-import net.spell_engine.api.spell.CustomSpellHandler;
+import net.spell_engine.api.spell.event.CustomSpellHandler;
 import net.spell_engine.api.spell.Spell;
+import net.spell_engine.api.spell.registry.SpellRegistry;
 import net.spell_engine.internals.SpellHelper;
 import net.spell_engine.utils.SoundHelper;
 import net.spell_engine.utils.TargetHelper;
-import net.spell_engine.api.spell.SpellInfo;
 
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Predicate;
 
 import static net.forcemaster_rpg.ForcemasterClassMod.MOD_ID;
-import static net.spell_engine.internals.SpellRegistry.getSpell;
 
 public class CustomSpells {
     public static void register() {
@@ -24,31 +23,30 @@ public class CustomSpells {
         int speed_belial_smashing = 4;
         double knockup_burstcrack = 0.35;
         double knockup_burstcrack_stonehand = 0.65;
-        int stun_duration_straight_punch = 60;
 
         /////FORCEMASTER_SPELLS
         /// BURSTCRACK
         CustomSpellHandler.register(Identifier.of(MOD_ID, "burstcrack"), (data) -> {
-            SpellInfo spellinfo = new SpellInfo(getSpell(Identifier.of(MOD_ID, "burstcrack")),Identifier.of(MOD_ID));
-            Spell.Impact[] impacts = getSpell(Identifier.of(MOD_ID, "burstcrack")).impact;
-
-
             CustomSpellHandler.Data data1 = (CustomSpellHandler.Data) data;
+            var spellEntry = SpellRegistry.from(data1.caster().getWorld()).getEntry(Identifier.of(MOD_ID, "burstcrack")).orElse(null);
+            var spell = spellEntry.value();
+            Spell.Impact[] impacts = spell.impact;
+
             Predicate<Entity> selectionPredicate = (target2) -> {
                 return (TargetHelper.actionAllowed(TargetHelper.TargetingMode.AREA, TargetHelper.Intent.HARMFUL, data1.caster(), target2)
                 );
             };
             if (!data1.caster().getWorld().isClient) {
-                List<Entity> list = data1.caster().getWorld().getOtherEntities(data1.caster(), data1.caster().getBoundingBox().expand(getSpell(Identifier.of(MOD_ID, "burstcrack")).range), selectionPredicate);
+                List<Entity> list = data1.caster().getWorld().getOtherEntities(data1.caster(), data1.caster().getBoundingBox().expand(spell.range), selectionPredicate);
                 for (Entity entity : list) {
-                    SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellinfo,impacts ,data1.impactContext());
-                    SoundHelper.playSound(data1.caster().getWorld(), entity, getSpell(Identifier.of(MOD_ID, "burstcrack")).impact[0].sound);
+                    SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellEntry,impacts ,data1.impactContext());
+                    SoundHelper.playSound(data1.caster().getWorld(), entity, spell.impact[0].sound);
                     Vec3d currentMovement = entity.getVelocity();
                     entity.setVelocity(currentMovement.x, currentMovement.y + knockup_burstcrack, currentMovement.z);
                     entity.velocityModified = true;
                     if (data1.caster().hasStatusEffect(Effects.STONE_HAND.registryEntry)) {
-                        SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellinfo,impacts ,data1.impactContext());
-                        SoundHelper.playSound(data1.caster().getWorld(), entity, getSpell(Identifier.of(MOD_ID, "burstcrack")).impact[0].sound);
+                        SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellEntry,impacts ,data1.impactContext());
+                        SoundHelper.playSound(data1.caster().getWorld(), entity, spell.impact[0].sound);
                         Vec3d currentMovement2 = entity.getVelocity();
                         entity.setVelocity(currentMovement2.x, currentMovement2.y + knockup_burstcrack_stonehand, currentMovement2.z);
                         entity.velocityModified = true;
@@ -60,10 +58,12 @@ public class CustomSpells {
         });
         /// BELIAL SMASHING
         CustomSpellHandler.register(Identifier.of(MOD_ID, "belial_smashing"), (data) -> {
-            SpellInfo spellinfo = new SpellInfo(getSpell(Identifier.of(MOD_ID, "belial_smashing")),Identifier.of(MOD_ID));
-            Spell.Impact[] impacts = getSpell(Identifier.of(MOD_ID, "belial_smashing")).impact;
             CustomSpellHandler.Data data1 = (CustomSpellHandler.Data) data;
-            List<Entity> list = TargetHelper.targetsFromRaycast(data1.caster(), getSpell(Identifier.of(MOD_ID, "belial_smashing")).range, Objects::nonNull);
+            var spellEntry = SpellRegistry.from(data1.caster().getWorld()).getEntry(Identifier.of(MOD_ID, "belial_smashing")).orElse(null);
+            var spell = spellEntry.value();
+            Spell.Impact[] impacts = spell.impact;
+
+            List<Entity> list = TargetHelper.targetsFromRaycast(data1.caster(), spell.range, Objects::nonNull);
             if (!data1.caster().getWorld().isClient) {
                 data1.caster().velocityDirty = true;
                 data1.caster().velocityModified = true;
@@ -71,7 +71,7 @@ public class CustomSpells {
                 Vec3d velocity = data1.caster().getVelocity();
                 data1.caster().addVelocity(rotationVector.x * 0.1 + (rotationVector.x * 2.5 - velocity.x) * speed_belial_smashing, 0, rotationVector.z * 0.1 + (rotationVector.z * 2.5 - velocity.z) * speed_belial_smashing);
                 for (Entity entity : list) {
-                    SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellinfo,impacts ,data1.impactContext());
+                    SpellHelper.performImpacts(entity.getWorld(), data1.caster(), entity, entity, spellEntry,impacts ,data1.impactContext());
                     Vec3d currentMovement2 = entity.getVelocity();
                     entity.setVelocity(currentMovement2.x, currentMovement2.y + 0.6f, currentMovement2.z);
                 }